OOP for thermophysical calculation.
Object-oriented programming was applied to calculate thermophysical
properties. In this way it is more easy to implement new models for
thermophysical properties. In order to check the code, results were
compared from previous code taking into accout Euler and Navier-Stokes
model with constant and variable thermophysical properties. In all cases
the results are exactly the same.